added live webcam preview

+++ src/main/java/parkman/Controllers/MainController.java
@@ -85,7 +85,9 @@ public class MainController implements Initializable {
             stage.setScene(new Scene(root, USE_COMPUTED_SIZE, USE_COMPUTED_SIZE));
             stage.setResizable(false);
             stage.showAndWait();
-            scanInDialogController.getImageFile().delete();
+
+            if(scanInDialogController.getImageFile() != null)
+                scanInDialogController.getImageFile().delete();
         } catch (IOException e) {
             e.printStackTrace();
         }
@@ -106,7 +108,9 @@ public class MainController implements Initializable {
             stage.setScene(new Scene(root, USE_COMPUTED_SIZE, USE_COMPUTED_SIZE));
             stage.setResizable(false);
             stage.showAndWait();
-            scanOutDialogController.getImageFile().delete();
+
+            if(scanOutDialogController.getImageFile() != null)
+                scanOutDialogController.getImageFile().delete();
         } catch (IOException e) {
             e.printStackTrace();
         }
